Fix error caused by events updating the component state not being wrapped in act()
The error:
Warning: An update to Homepage inside a test was not wrapped in act(...).
When testing, code that causes React state updates should be wrapped into act(...):

act(() => {
    /* fire events that update state */
});
/* assert on the output */

This ensures that you're testing the behavior the user would see in the browser. Learn more at https://reactjs.org/link/wrap-tests-with-act